+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- `.authenticated`
+- `.repositories.getUserRepository()`
+- `.repositories.createUserRepository()`
+- Throws error on calls if not authenticated
 - API documentation
 
 ### Changed